Bind Multiple Track Tables by Row

bind_track_df and rbind.track_df are modified versions of dplyr::bind_rows and base::rbind respectively, that work with track tables. In particular, they make sure that you cannot bind together two tables with different projections and, in the case of rbind.track_df, that the projection attribute is inherited by the resulting track table (dplyr::bind_rows strips it otherwise).

bind_track_df(..., .id = NULL)

# S3 method for track_df
rbind(...)

Arguments

...
Track tables to combine.

Each argument can either be a track table, a list that could be track table, or a list of track table. Columns are matched by name, and any missing columns will be filled with NA.

.id
Track table identifier.

When .id is supplied, a new column of identifiers is created to link each row to its original data frame. The labels are taken from the named arguments to bind_rows(). When a list of data frames is supplied, the labels are taken from the names of the list. If no names are found a numeric sequence is used instead.

Value

A track table.

Note

dplyr::bind_rows cannot be converted to a S3 generic and therefore cannot implement convenient dispatching based on the class of the tables it is provided with. As a consequence, it is not possible at the moment to provide specific implementations of bind_row for track tables.

Examples

# TODO