Bind Multiple Track Tables by Row

bind_tracks uses data.table::rbindlist to combine track tables by rows, but makes sure that you cannot bind together two tables with different projections or time zones, that the projection attribute is inherited by the resulting track table, and that track tables based on different table classes are coerced to the same table class.

Usage

bind_tracks(...)

Arguments

...

A list containing track table objects, or the names of track table objects separated by commas. The track tables must have the same projection and time zone.

Value

A track table.

Author

Simon Garnier, garnier@njit.edu

Examples

data(short_tracks)

bind_tracks(short_tracks, short_tracks)
#> Track table [240 observations]
#> Number of tracks:  2 
#> Dimensions:  2D 
#> Geographic:  TRUE 
#> Projection:  +proj=longlat 
#> Table class:  data frame ('data.frame')
#>     id                   t        x         y
#> 1    1 2015-09-10 07:01:00 15.76468 -22.37961
#> 2    1 2015-09-10 07:01:01 15.76469 -22.37960
#> 3    1 2015-09-10 07:01:02 15.76469 -22.37960
#> 4    1 2015-09-10 07:01:03 15.76469 -22.37960
#> 5    1 2015-09-10 07:01:04 15.76470 -22.37960
#> 6    1 2015-09-10 07:01:05 15.76469 -22.37960
#> 7    1 2015-09-10 07:01:06 15.76469 -22.37960
#> 8    1 2015-09-10 07:01:07 15.76469 -22.37959
#> 9    1 2015-09-10 07:01:08 15.76469 -22.37959
#> 10   1 2015-09-10 07:01:09 15.76469 -22.37959
#> 11   1 2015-09-10 07:01:10 15.76469 -22.37959
#> 12   1 2015-09-10 07:01:11 15.76469 -22.37960
#> 13   1 2015-09-10 07:01:12 15.76469 -22.37960
#> 14   1 2015-09-10 07:01:13 15.76469 -22.37960
#> 15   1 2015-09-10 07:01:14 15.76469 -22.37960
#> 16   1 2015-09-10 07:01:15 15.76468 -22.37960
#> 17   1 2015-09-10 07:01:16 15.76468 -22.37960
#> 18   1 2015-09-10 07:01:17 15.76468 -22.37960
#> 19   1 2015-09-10 07:01:18 15.76468 -22.37960
#> 20   1 2015-09-10 07:01:19 15.76468 -22.37960
#> 21   1 2015-09-10 07:01:20 15.76468 -22.37960
#> 22   1 2015-09-10 07:01:21 15.76468 -22.37960
#> 23   1 2015-09-10 07:01:22 15.76468 -22.37960
#> 24   1 2015-09-10 07:01:23 15.76468 -22.37960
#> 25   1 2015-09-10 07:01:24 15.76468 -22.37960
#> 26   1 2015-09-10 07:01:25 15.76468 -22.37960
#> 27   1 2015-09-10 07:01:26 15.76468 -22.37960
#> 28   1 2015-09-10 07:01:27 15.76468 -22.37960
#> 29   1 2015-09-10 07:01:28 15.76468 -22.37960
#> 30   1 2015-09-10 07:01:29 15.76468 -22.37960
#> 31   1 2015-09-10 07:01:30 15.76468 -22.37960
#> 32   1 2015-09-10 07:01:31 15.76468 -22.37960
#> 33   1 2015-09-10 07:01:32 15.76469 -22.37960
#> 34   1 2015-09-10 07:01:33 15.76469 -22.37960
#> 35   1 2015-09-10 07:01:34 15.76469 -22.37960
#> 36   1 2015-09-10 07:01:35 15.76469 -22.37960
#> 37   1 2015-09-10 07:01:36 15.76469 -22.37960
#> 38   1 2015-09-10 07:01:37 15.76469 -22.37960
#> 39   1 2015-09-10 07:01:38 15.76469 -22.37960
#> 40   1 2015-09-10 07:01:39 15.76469 -22.37960
#> 41   1 2015-09-10 07:01:40 15.76469 -22.37960
#> 42   1 2015-09-10 07:01:41 15.76469 -22.37960
#> 43   1 2015-09-10 07:01:42 15.76469 -22.37960
#> 44   1 2015-09-10 07:01:43 15.76469 -22.37960
#> 45   1 2015-09-10 07:01:44 15.76469 -22.37960
#> 46   1 2015-09-10 07:01:45 15.76469 -22.37960
#> 47   1 2015-09-10 07:01:46 15.76469 -22.37960
#> 48   1 2015-09-10 07:01:47 15.76469 -22.37960
#> 49   1 2015-09-10 07:01:48 15.76469 -22.37960
#> 50   1 2015-09-10 07:01:49 15.76469 -22.37959
#> 51   1 2015-09-10 07:01:50 15.76469 -22.37959
#> 52   1 2015-09-10 07:01:51 15.76470 -22.37959
#> 53   1 2015-09-10 07:01:52 15.76470 -22.37959
#> 54   1 2015-09-10 07:01:53 15.76470 -22.37959
#> 55   1 2015-09-10 07:01:54 15.76470 -22.37960
#> 56   1 2015-09-10 07:01:55 15.76470 -22.37960
#> 57   1 2015-09-10 07:01:56 15.76471 -22.37960
#> 58   1 2015-09-10 07:01:57 15.76471 -22.37959
#> 59   1 2015-09-10 07:01:58 15.76471 -22.37959
#> 60   1 2015-09-10 07:01:59 15.76471 -22.37959
#> 61   2 2015-09-10 07:01:00 15.76459 -22.37971
#> 62   2 2015-09-10 07:01:01 15.76459 -22.37971
#> 63   2 2015-09-10 07:01:02 15.76459 -22.37971
#> 64   2 2015-09-10 07:01:03 15.76459 -22.37971
#> 65   2 2015-09-10 07:01:04 15.76459 -22.37971
#> 66   2 2015-09-10 07:01:05 15.76459 -22.37971
#> 67   2 2015-09-10 07:01:06 15.76459 -22.37971
#> 68   2 2015-09-10 07:01:07 15.76459 -22.37971
#> 69   2 2015-09-10 07:01:08 15.76459 -22.37971
#> 70   2 2015-09-10 07:01:09 15.76459 -22.37971
#> 71   2 2015-09-10 07:01:10 15.76459 -22.37971
#> 72   2 2015-09-10 07:01:11 15.76459 -22.37971
#> 73   2 2015-09-10 07:01:12 15.76459 -22.37971
#> 74   2 2015-09-10 07:01:13 15.76459 -22.37971
#> 75   2 2015-09-10 07:01:14 15.76459 -22.37971
#> 76   2 2015-09-10 07:01:15 15.76459 -22.37971
#> 77   2 2015-09-10 07:01:16 15.76459 -22.37971
#> 78   2 2015-09-10 07:01:17 15.76459 -22.37971
#> 79   2 2015-09-10 07:01:18 15.76459 -22.37971
#> 80   2 2015-09-10 07:01:19 15.76459 -22.37971
#> 81   2 2015-09-10 07:01:20 15.76459 -22.37971
#> 82   2 2015-09-10 07:01:21 15.76459 -22.37971
#> 83   2 2015-09-10 07:01:22 15.76459 -22.37971
#> 84   2 2015-09-10 07:01:23 15.76459 -22.37971
#> 85   2 2015-09-10 07:01:24 15.76459 -22.37971
#> 86   2 2015-09-10 07:01:25 15.76458 -22.37971
#> 87   2 2015-09-10 07:01:26 15.76459 -22.37971
#> 88   2 2015-09-10 07:01:27 15.76459 -22.37971
#> 89   2 2015-09-10 07:01:28 15.76459 -22.37971
#> 90   2 2015-09-10 07:01:29 15.76458 -22.37971
#> 91   2 2015-09-10 07:01:30 15.76458 -22.37971
#> 92   2 2015-09-10 07:01:31 15.76457 -22.37971
#> 93   2 2015-09-10 07:01:32 15.76456 -22.37971
#> 94   2 2015-09-10 07:01:33 15.76455 -22.37972
#> 95   2 2015-09-10 07:01:34 15.76455 -22.37972
#> 96   2 2015-09-10 07:01:35 15.76455 -22.37973
#> 97   2 2015-09-10 07:01:36 15.76455 -22.37973
#> 98   2 2015-09-10 07:01:37 15.76455 -22.37974
#> 99   2 2015-09-10 07:01:38 15.76455 -22.37974
#> 100  2 2015-09-10 07:01:39 15.76455 -22.37975
#> 101  2 2015-09-10 07:01:40 15.76455 -22.37975
#> 102  2 2015-09-10 07:01:41 15.76454 -22.37976
#> 103  2 2015-09-10 07:01:42 15.76455 -22.37976
#> 104  2 2015-09-10 07:01:43 15.76455 -22.37976
#> 105  2 2015-09-10 07:01:44 15.76455 -22.37977
#> 106  2 2015-09-10 07:01:45 15.76456 -22.37977
#> 107  2 2015-09-10 07:01:46 15.76457 -22.37978
#> 108  2 2015-09-10 07:01:47 15.76458 -22.37978
#> 109  2 2015-09-10 07:01:48 15.76459 -22.37979
#> 110  2 2015-09-10 07:01:49 15.76459 -22.37979
#> 111  2 2015-09-10 07:01:50 15.76460 -22.37980
#> 112  2 2015-09-10 07:01:51 15.76461 -22.37980
#> 113  2 2015-09-10 07:01:52 15.76461 -22.37981
#> 114  2 2015-09-10 07:01:53 15.76461 -22.37981
#> 115  2 2015-09-10 07:01:54 15.76462 -22.37981
#> 116  2 2015-09-10 07:01:55 15.76463 -22.37982
#> 117  2 2015-09-10 07:01:56 15.76463 -22.37982
#> 118  2 2015-09-10 07:01:57 15.76464 -22.37982
#> 119  2 2015-09-10 07:01:58 15.76465 -22.37982
#> 120  2 2015-09-10 07:01:59 15.76466 -22.37982
#> 121  1 2015-09-10 07:01:00 15.76468 -22.37961
#> 122  1 2015-09-10 07:01:01 15.76469 -22.37960
#> 123  1 2015-09-10 07:01:02 15.76469 -22.37960
#> 124  1 2015-09-10 07:01:03 15.76469 -22.37960
#> 125  1 2015-09-10 07:01:04 15.76470 -22.37960
#> 126  1 2015-09-10 07:01:05 15.76469 -22.37960
#> 127  1 2015-09-10 07:01:06 15.76469 -22.37960
#> 128  1 2015-09-10 07:01:07 15.76469 -22.37959
#> 129  1 2015-09-10 07:01:08 15.76469 -22.37959
#> 130  1 2015-09-10 07:01:09 15.76469 -22.37959
#> 131  1 2015-09-10 07:01:10 15.76469 -22.37959
#> 132  1 2015-09-10 07:01:11 15.76469 -22.37960
#> 133  1 2015-09-10 07:01:12 15.76469 -22.37960
#> 134  1 2015-09-10 07:01:13 15.76469 -22.37960
#> 135  1 2015-09-10 07:01:14 15.76469 -22.37960
#> 136  1 2015-09-10 07:01:15 15.76468 -22.37960
#> 137  1 2015-09-10 07:01:16 15.76468 -22.37960
#> 138  1 2015-09-10 07:01:17 15.76468 -22.37960
#> 139  1 2015-09-10 07:01:18 15.76468 -22.37960
#> 140  1 2015-09-10 07:01:19 15.76468 -22.37960
#> 141  1 2015-09-10 07:01:20 15.76468 -22.37960
#> 142  1 2015-09-10 07:01:21 15.76468 -22.37960
#> 143  1 2015-09-10 07:01:22 15.76468 -22.37960
#> 144  1 2015-09-10 07:01:23 15.76468 -22.37960
#> 145  1 2015-09-10 07:01:24 15.76468 -22.37960
#> 146  1 2015-09-10 07:01:25 15.76468 -22.37960
#> 147  1 2015-09-10 07:01:26 15.76468 -22.37960
#> 148  1 2015-09-10 07:01:27 15.76468 -22.37960
#> 149  1 2015-09-10 07:01:28 15.76468 -22.37960
#> 150  1 2015-09-10 07:01:29 15.76468 -22.37960
#> 151  1 2015-09-10 07:01:30 15.76468 -22.37960
#> 152  1 2015-09-10 07:01:31 15.76468 -22.37960
#> 153  1 2015-09-10 07:01:32 15.76469 -22.37960
#> 154  1 2015-09-10 07:01:33 15.76469 -22.37960
#> 155  1 2015-09-10 07:01:34 15.76469 -22.37960
#> 156  1 2015-09-10 07:01:35 15.76469 -22.37960
#> 157  1 2015-09-10 07:01:36 15.76469 -22.37960
#> 158  1 2015-09-10 07:01:37 15.76469 -22.37960
#> 159  1 2015-09-10 07:01:38 15.76469 -22.37960
#> 160  1 2015-09-10 07:01:39 15.76469 -22.37960
#> 161  1 2015-09-10 07:01:40 15.76469 -22.37960
#> 162  1 2015-09-10 07:01:41 15.76469 -22.37960
#> 163  1 2015-09-10 07:01:42 15.76469 -22.37960
#> 164  1 2015-09-10 07:01:43 15.76469 -22.37960
#> 165  1 2015-09-10 07:01:44 15.76469 -22.37960
#> 166  1 2015-09-10 07:01:45 15.76469 -22.37960
#> 167  1 2015-09-10 07:01:46 15.76469 -22.37960
#> 168  1 2015-09-10 07:01:47 15.76469 -22.37960
#> 169  1 2015-09-10 07:01:48 15.76469 -22.37960
#> 170  1 2015-09-10 07:01:49 15.76469 -22.37959
#> 171  1 2015-09-10 07:01:50 15.76469 -22.37959
#> 172  1 2015-09-10 07:01:51 15.76470 -22.37959
#> 173  1 2015-09-10 07:01:52 15.76470 -22.37959
#> 174  1 2015-09-10 07:01:53 15.76470 -22.37959
#> 175  1 2015-09-10 07:01:54 15.76470 -22.37960
#> 176  1 2015-09-10 07:01:55 15.76470 -22.37960
#> 177  1 2015-09-10 07:01:56 15.76471 -22.37960
#> 178  1 2015-09-10 07:01:57 15.76471 -22.37959
#> 179  1 2015-09-10 07:01:58 15.76471 -22.37959
#> 180  1 2015-09-10 07:01:59 15.76471 -22.37959
#> 181  2 2015-09-10 07:01:00 15.76459 -22.37971
#> 182  2 2015-09-10 07:01:01 15.76459 -22.37971
#> 183  2 2015-09-10 07:01:02 15.76459 -22.37971
#> 184  2 2015-09-10 07:01:03 15.76459 -22.37971
#> 185  2 2015-09-10 07:01:04 15.76459 -22.37971
#> 186  2 2015-09-10 07:01:05 15.76459 -22.37971
#> 187  2 2015-09-10 07:01:06 15.76459 -22.37971
#> 188  2 2015-09-10 07:01:07 15.76459 -22.37971
#> 189  2 2015-09-10 07:01:08 15.76459 -22.37971
#> 190  2 2015-09-10 07:01:09 15.76459 -22.37971
#> 191  2 2015-09-10 07:01:10 15.76459 -22.37971
#> 192  2 2015-09-10 07:01:11 15.76459 -22.37971
#> 193  2 2015-09-10 07:01:12 15.76459 -22.37971
#> 194  2 2015-09-10 07:01:13 15.76459 -22.37971
#> 195  2 2015-09-10 07:01:14 15.76459 -22.37971
#> 196  2 2015-09-10 07:01:15 15.76459 -22.37971
#> 197  2 2015-09-10 07:01:16 15.76459 -22.37971
#> 198  2 2015-09-10 07:01:17 15.76459 -22.37971
#> 199  2 2015-09-10 07:01:18 15.76459 -22.37971
#> 200  2 2015-09-10 07:01:19 15.76459 -22.37971
#> 201  2 2015-09-10 07:01:20 15.76459 -22.37971
#> 202  2 2015-09-10 07:01:21 15.76459 -22.37971
#> 203  2 2015-09-10 07:01:22 15.76459 -22.37971
#> 204  2 2015-09-10 07:01:23 15.76459 -22.37971
#> 205  2 2015-09-10 07:01:24 15.76459 -22.37971
#> 206  2 2015-09-10 07:01:25 15.76458 -22.37971
#> 207  2 2015-09-10 07:01:26 15.76459 -22.37971
#> 208  2 2015-09-10 07:01:27 15.76459 -22.37971
#> 209  2 2015-09-10 07:01:28 15.76459 -22.37971
#> 210  2 2015-09-10 07:01:29 15.76458 -22.37971
#> 211  2 2015-09-10 07:01:30 15.76458 -22.37971
#> 212  2 2015-09-10 07:01:31 15.76457 -22.37971
#> 213  2 2015-09-10 07:01:32 15.76456 -22.37971
#> 214  2 2015-09-10 07:01:33 15.76455 -22.37972
#> 215  2 2015-09-10 07:01:34 15.76455 -22.37972
#> 216  2 2015-09-10 07:01:35 15.76455 -22.37973
#> 217  2 2015-09-10 07:01:36 15.76455 -22.37973
#> 218  2 2015-09-10 07:01:37 15.76455 -22.37974
#> 219  2 2015-09-10 07:01:38 15.76455 -22.37974
#> 220  2 2015-09-10 07:01:39 15.76455 -22.37975
#> 221  2 2015-09-10 07:01:40 15.76455 -22.37975
#> 222  2 2015-09-10 07:01:41 15.76454 -22.37976
#> 223  2 2015-09-10 07:01:42 15.76455 -22.37976
#> 224  2 2015-09-10 07:01:43 15.76455 -22.37976
#> 225  2 2015-09-10 07:01:44 15.76455 -22.37977
#> 226  2 2015-09-10 07:01:45 15.76456 -22.37977
#> 227  2 2015-09-10 07:01:46 15.76457 -22.37978
#> 228  2 2015-09-10 07:01:47 15.76458 -22.37978
#> 229  2 2015-09-10 07:01:48 15.76459 -22.37979
#> 230  2 2015-09-10 07:01:49 15.76459 -22.37979
#> 231  2 2015-09-10 07:01:50 15.76460 -22.37980
#> 232  2 2015-09-10 07:01:51 15.76461 -22.37980
#> 233  2 2015-09-10 07:01:52 15.76461 -22.37981
#> 234  2 2015-09-10 07:01:53 15.76461 -22.37981
#> 235  2 2015-09-10 07:01:54 15.76462 -22.37981
#> 236  2 2015-09-10 07:01:55 15.76463 -22.37982
#> 237  2 2015-09-10 07:01:56 15.76463 -22.37982
#> 238  2 2015-09-10 07:01:57 15.76464 -22.37982
#> 239  2 2015-09-10 07:01:58 15.76465 -22.37982
#> 240  2 2015-09-10 07:01:59 15.76466 -22.37982
bind_tracks(list(short_tracks, short_tracks))
#> Track table [240 observations]
#> Number of tracks:  2 
#> Dimensions:  2D 
#> Geographic:  TRUE 
#> Projection:  +proj=longlat 
#> Table class:  data frame ('data.frame')
#>     id                   t        x         y
#> 1    1 2015-09-10 07:01:00 15.76468 -22.37961
#> 2    1 2015-09-10 07:01:01 15.76469 -22.37960
#> 3    1 2015-09-10 07:01:02 15.76469 -22.37960
#> 4    1 2015-09-10 07:01:03 15.76469 -22.37960
#> 5    1 2015-09-10 07:01:04 15.76470 -22.37960
#> 6    1 2015-09-10 07:01:05 15.76469 -22.37960
#> 7    1 2015-09-10 07:01:06 15.76469 -22.37960
#> 8    1 2015-09-10 07:01:07 15.76469 -22.37959
#> 9    1 2015-09-10 07:01:08 15.76469 -22.37959
#> 10   1 2015-09-10 07:01:09 15.76469 -22.37959
#> 11   1 2015-09-10 07:01:10 15.76469 -22.37959
#> 12   1 2015-09-10 07:01:11 15.76469 -22.37960
#> 13   1 2015-09-10 07:01:12 15.76469 -22.37960
#> 14   1 2015-09-10 07:01:13 15.76469 -22.37960
#> 15   1 2015-09-10 07:01:14 15.76469 -22.37960
#> 16   1 2015-09-10 07:01:15 15.76468 -22.37960
#> 17   1 2015-09-10 07:01:16 15.76468 -22.37960
#> 18   1 2015-09-10 07:01:17 15.76468 -22.37960
#> 19   1 2015-09-10 07:01:18 15.76468 -22.37960
#> 20   1 2015-09-10 07:01:19 15.76468 -22.37960
#> 21   1 2015-09-10 07:01:20 15.76468 -22.37960
#> 22   1 2015-09-10 07:01:21 15.76468 -22.37960
#> 23   1 2015-09-10 07:01:22 15.76468 -22.37960
#> 24   1 2015-09-10 07:01:23 15.76468 -22.37960
#> 25   1 2015-09-10 07:01:24 15.76468 -22.37960
#> 26   1 2015-09-10 07:01:25 15.76468 -22.37960
#> 27   1 2015-09-10 07:01:26 15.76468 -22.37960
#> 28   1 2015-09-10 07:01:27 15.76468 -22.37960
#> 29   1 2015-09-10 07:01:28 15.76468 -22.37960
#> 30   1 2015-09-10 07:01:29 15.76468 -22.37960
#> 31   1 2015-09-10 07:01:30 15.76468 -22.37960
#> 32   1 2015-09-10 07:01:31 15.76468 -22.37960
#> 33   1 2015-09-10 07:01:32 15.76469 -22.37960
#> 34   1 2015-09-10 07:01:33 15.76469 -22.37960
#> 35   1 2015-09-10 07:01:34 15.76469 -22.37960
#> 36   1 2015-09-10 07:01:35 15.76469 -22.37960
#> 37   1 2015-09-10 07:01:36 15.76469 -22.37960
#> 38   1 2015-09-10 07:01:37 15.76469 -22.37960
#> 39   1 2015-09-10 07:01:38 15.76469 -22.37960
#> 40   1 2015-09-10 07:01:39 15.76469 -22.37960
#> 41   1 2015-09-10 07:01:40 15.76469 -22.37960
#> 42   1 2015-09-10 07:01:41 15.76469 -22.37960
#> 43   1 2015-09-10 07:01:42 15.76469 -22.37960
#> 44   1 2015-09-10 07:01:43 15.76469 -22.37960
#> 45   1 2015-09-10 07:01:44 15.76469 -22.37960
#> 46   1 2015-09-10 07:01:45 15.76469 -22.37960
#> 47   1 2015-09-10 07:01:46 15.76469 -22.37960
#> 48   1 2015-09-10 07:01:47 15.76469 -22.37960
#> 49   1 2015-09-10 07:01:48 15.76469 -22.37960
#> 50   1 2015-09-10 07:01:49 15.76469 -22.37959
#> 51   1 2015-09-10 07:01:50 15.76469 -22.37959
#> 52   1 2015-09-10 07:01:51 15.76470 -22.37959
#> 53   1 2015-09-10 07:01:52 15.76470 -22.37959
#> 54   1 2015-09-10 07:01:53 15.76470 -22.37959
#> 55   1 2015-09-10 07:01:54 15.76470 -22.37960
#> 56   1 2015-09-10 07:01:55 15.76470 -22.37960
#> 57   1 2015-09-10 07:01:56 15.76471 -22.37960
#> 58   1 2015-09-10 07:01:57 15.76471 -22.37959
#> 59   1 2015-09-10 07:01:58 15.76471 -22.37959
#> 60   1 2015-09-10 07:01:59 15.76471 -22.37959
#> 61   2 2015-09-10 07:01:00 15.76459 -22.37971
#> 62   2 2015-09-10 07:01:01 15.76459 -22.37971
#> 63   2 2015-09-10 07:01:02 15.76459 -22.37971
#> 64   2 2015-09-10 07:01:03 15.76459 -22.37971
#> 65   2 2015-09-10 07:01:04 15.76459 -22.37971
#> 66   2 2015-09-10 07:01:05 15.76459 -22.37971
#> 67   2 2015-09-10 07:01:06 15.76459 -22.37971
#> 68   2 2015-09-10 07:01:07 15.76459 -22.37971
#> 69   2 2015-09-10 07:01:08 15.76459 -22.37971
#> 70   2 2015-09-10 07:01:09 15.76459 -22.37971
#> 71   2 2015-09-10 07:01:10 15.76459 -22.37971
#> 72   2 2015-09-10 07:01:11 15.76459 -22.37971
#> 73   2 2015-09-10 07:01:12 15.76459 -22.37971
#> 74   2 2015-09-10 07:01:13 15.76459 -22.37971
#> 75   2 2015-09-10 07:01:14 15.76459 -22.37971
#> 76   2 2015-09-10 07:01:15 15.76459 -22.37971
#> 77   2 2015-09-10 07:01:16 15.76459 -22.37971
#> 78   2 2015-09-10 07:01:17 15.76459 -22.37971
#> 79   2 2015-09-10 07:01:18 15.76459 -22.37971
#> 80   2 2015-09-10 07:01:19 15.76459 -22.37971
#> 81   2 2015-09-10 07:01:20 15.76459 -22.37971
#> 82   2 2015-09-10 07:01:21 15.76459 -22.37971
#> 83   2 2015-09-10 07:01:22 15.76459 -22.37971
#> 84   2 2015-09-10 07:01:23 15.76459 -22.37971
#> 85   2 2015-09-10 07:01:24 15.76459 -22.37971
#> 86   2 2015-09-10 07:01:25 15.76458 -22.37971
#> 87   2 2015-09-10 07:01:26 15.76459 -22.37971
#> 88   2 2015-09-10 07:01:27 15.76459 -22.37971
#> 89   2 2015-09-10 07:01:28 15.76459 -22.37971
#> 90   2 2015-09-10 07:01:29 15.76458 -22.37971
#> 91   2 2015-09-10 07:01:30 15.76458 -22.37971
#> 92   2 2015-09-10 07:01:31 15.76457 -22.37971
#> 93   2 2015-09-10 07:01:32 15.76456 -22.37971
#> 94   2 2015-09-10 07:01:33 15.76455 -22.37972
#> 95   2 2015-09-10 07:01:34 15.76455 -22.37972
#> 96   2 2015-09-10 07:01:35 15.76455 -22.37973
#> 97   2 2015-09-10 07:01:36 15.76455 -22.37973
#> 98   2 2015-09-10 07:01:37 15.76455 -22.37974
#> 99   2 2015-09-10 07:01:38 15.76455 -22.37974
#> 100  2 2015-09-10 07:01:39 15.76455 -22.37975
#> 101  2 2015-09-10 07:01:40 15.76455 -22.37975
#> 102  2 2015-09-10 07:01:41 15.76454 -22.37976
#> 103  2 2015-09-10 07:01:42 15.76455 -22.37976
#> 104  2 2015-09-10 07:01:43 15.76455 -22.37976
#> 105  2 2015-09-10 07:01:44 15.76455 -22.37977
#> 106  2 2015-09-10 07:01:45 15.76456 -22.37977
#> 107  2 2015-09-10 07:01:46 15.76457 -22.37978
#> 108  2 2015-09-10 07:01:47 15.76458 -22.37978
#> 109  2 2015-09-10 07:01:48 15.76459 -22.37979
#> 110  2 2015-09-10 07:01:49 15.76459 -22.37979
#> 111  2 2015-09-10 07:01:50 15.76460 -22.37980
#> 112  2 2015-09-10 07:01:51 15.76461 -22.37980
#> 113  2 2015-09-10 07:01:52 15.76461 -22.37981
#> 114  2 2015-09-10 07:01:53 15.76461 -22.37981
#> 115  2 2015-09-10 07:01:54 15.76462 -22.37981
#> 116  2 2015-09-10 07:01:55 15.76463 -22.37982
#> 117  2 2015-09-10 07:01:56 15.76463 -22.37982
#> 118  2 2015-09-10 07:01:57 15.76464 -22.37982
#> 119  2 2015-09-10 07:01:58 15.76465 -22.37982
#> 120  2 2015-09-10 07:01:59 15.76466 -22.37982
#> 121  1 2015-09-10 07:01:00 15.76468 -22.37961
#> 122  1 2015-09-10 07:01:01 15.76469 -22.37960
#> 123  1 2015-09-10 07:01:02 15.76469 -22.37960
#> 124  1 2015-09-10 07:01:03 15.76469 -22.37960
#> 125  1 2015-09-10 07:01:04 15.76470 -22.37960
#> 126  1 2015-09-10 07:01:05 15.76469 -22.37960
#> 127  1 2015-09-10 07:01:06 15.76469 -22.37960
#> 128  1 2015-09-10 07:01:07 15.76469 -22.37959
#> 129  1 2015-09-10 07:01:08 15.76469 -22.37959
#> 130  1 2015-09-10 07:01:09 15.76469 -22.37959
#> 131  1 2015-09-10 07:01:10 15.76469 -22.37959
#> 132  1 2015-09-10 07:01:11 15.76469 -22.37960
#> 133  1 2015-09-10 07:01:12 15.76469 -22.37960
#> 134  1 2015-09-10 07:01:13 15.76469 -22.37960
#> 135  1 2015-09-10 07:01:14 15.76469 -22.37960
#> 136  1 2015-09-10 07:01:15 15.76468 -22.37960
#> 137  1 2015-09-10 07:01:16 15.76468 -22.37960
#> 138  1 2015-09-10 07:01:17 15.76468 -22.37960
#> 139  1 2015-09-10 07:01:18 15.76468 -22.37960
#> 140  1 2015-09-10 07:01:19 15.76468 -22.37960
#> 141  1 2015-09-10 07:01:20 15.76468 -22.37960
#> 142  1 2015-09-10 07:01:21 15.76468 -22.37960
#> 143  1 2015-09-10 07:01:22 15.76468 -22.37960
#> 144  1 2015-09-10 07:01:23 15.76468 -22.37960
#> 145  1 2015-09-10 07:01:24 15.76468 -22.37960
#> 146  1 2015-09-10 07:01:25 15.76468 -22.37960
#> 147  1 2015-09-10 07:01:26 15.76468 -22.37960
#> 148  1 2015-09-10 07:01:27 15.76468 -22.37960
#> 149  1 2015-09-10 07:01:28 15.76468 -22.37960
#> 150  1 2015-09-10 07:01:29 15.76468 -22.37960
#> 151  1 2015-09-10 07:01:30 15.76468 -22.37960
#> 152  1 2015-09-10 07:01:31 15.76468 -22.37960
#> 153  1 2015-09-10 07:01:32 15.76469 -22.37960
#> 154  1 2015-09-10 07:01:33 15.76469 -22.37960
#> 155  1 2015-09-10 07:01:34 15.76469 -22.37960
#> 156  1 2015-09-10 07:01:35 15.76469 -22.37960
#> 157  1 2015-09-10 07:01:36 15.76469 -22.37960
#> 158  1 2015-09-10 07:01:37 15.76469 -22.37960
#> 159  1 2015-09-10 07:01:38 15.76469 -22.37960
#> 160  1 2015-09-10 07:01:39 15.76469 -22.37960
#> 161  1 2015-09-10 07:01:40 15.76469 -22.37960
#> 162  1 2015-09-10 07:01:41 15.76469 -22.37960
#> 163  1 2015-09-10 07:01:42 15.76469 -22.37960
#> 164  1 2015-09-10 07:01:43 15.76469 -22.37960
#> 165  1 2015-09-10 07:01:44 15.76469 -22.37960
#> 166  1 2015-09-10 07:01:45 15.76469 -22.37960
#> 167  1 2015-09-10 07:01:46 15.76469 -22.37960
#> 168  1 2015-09-10 07:01:47 15.76469 -22.37960
#> 169  1 2015-09-10 07:01:48 15.76469 -22.37960
#> 170  1 2015-09-10 07:01:49 15.76469 -22.37959
#> 171  1 2015-09-10 07:01:50 15.76469 -22.37959
#> 172  1 2015-09-10 07:01:51 15.76470 -22.37959
#> 173  1 2015-09-10 07:01:52 15.76470 -22.37959
#> 174  1 2015-09-10 07:01:53 15.76470 -22.37959
#> 175  1 2015-09-10 07:01:54 15.76470 -22.37960
#> 176  1 2015-09-10 07:01:55 15.76470 -22.37960
#> 177  1 2015-09-10 07:01:56 15.76471 -22.37960
#> 178  1 2015-09-10 07:01:57 15.76471 -22.37959
#> 179  1 2015-09-10 07:01:58 15.76471 -22.37959
#> 180  1 2015-09-10 07:01:59 15.76471 -22.37959
#> 181  2 2015-09-10 07:01:00 15.76459 -22.37971
#> 182  2 2015-09-10 07:01:01 15.76459 -22.37971
#> 183  2 2015-09-10 07:01:02 15.76459 -22.37971
#> 184  2 2015-09-10 07:01:03 15.76459 -22.37971
#> 185  2 2015-09-10 07:01:04 15.76459 -22.37971
#> 186  2 2015-09-10 07:01:05 15.76459 -22.37971
#> 187  2 2015-09-10 07:01:06 15.76459 -22.37971
#> 188  2 2015-09-10 07:01:07 15.76459 -22.37971
#> 189  2 2015-09-10 07:01:08 15.76459 -22.37971
#> 190  2 2015-09-10 07:01:09 15.76459 -22.37971
#> 191  2 2015-09-10 07:01:10 15.76459 -22.37971
#> 192  2 2015-09-10 07:01:11 15.76459 -22.37971
#> 193  2 2015-09-10 07:01:12 15.76459 -22.37971
#> 194  2 2015-09-10 07:01:13 15.76459 -22.37971
#> 195  2 2015-09-10 07:01:14 15.76459 -22.37971
#> 196  2 2015-09-10 07:01:15 15.76459 -22.37971
#> 197  2 2015-09-10 07:01:16 15.76459 -22.37971
#> 198  2 2015-09-10 07:01:17 15.76459 -22.37971
#> 199  2 2015-09-10 07:01:18 15.76459 -22.37971
#> 200  2 2015-09-10 07:01:19 15.76459 -22.37971
#> 201  2 2015-09-10 07:01:20 15.76459 -22.37971
#> 202  2 2015-09-10 07:01:21 15.76459 -22.37971
#> 203  2 2015-09-10 07:01:22 15.76459 -22.37971
#> 204  2 2015-09-10 07:01:23 15.76459 -22.37971
#> 205  2 2015-09-10 07:01:24 15.76459 -22.37971
#> 206  2 2015-09-10 07:01:25 15.76458 -22.37971
#> 207  2 2015-09-10 07:01:26 15.76459 -22.37971
#> 208  2 2015-09-10 07:01:27 15.76459 -22.37971
#> 209  2 2015-09-10 07:01:28 15.76459 -22.37971
#> 210  2 2015-09-10 07:01:29 15.76458 -22.37971
#> 211  2 2015-09-10 07:01:30 15.76458 -22.37971
#> 212  2 2015-09-10 07:01:31 15.76457 -22.37971
#> 213  2 2015-09-10 07:01:32 15.76456 -22.37971
#> 214  2 2015-09-10 07:01:33 15.76455 -22.37972
#> 215  2 2015-09-10 07:01:34 15.76455 -22.37972
#> 216  2 2015-09-10 07:01:35 15.76455 -22.37973
#> 217  2 2015-09-10 07:01:36 15.76455 -22.37973
#> 218  2 2015-09-10 07:01:37 15.76455 -22.37974
#> 219  2 2015-09-10 07:01:38 15.76455 -22.37974
#> 220  2 2015-09-10 07:01:39 15.76455 -22.37975
#> 221  2 2015-09-10 07:01:40 15.76455 -22.37975
#> 222  2 2015-09-10 07:01:41 15.76454 -22.37976
#> 223  2 2015-09-10 07:01:42 15.76455 -22.37976
#> 224  2 2015-09-10 07:01:43 15.76455 -22.37976
#> 225  2 2015-09-10 07:01:44 15.76455 -22.37977
#> 226  2 2015-09-10 07:01:45 15.76456 -22.37977
#> 227  2 2015-09-10 07:01:46 15.76457 -22.37978
#> 228  2 2015-09-10 07:01:47 15.76458 -22.37978
#> 229  2 2015-09-10 07:01:48 15.76459 -22.37979
#> 230  2 2015-09-10 07:01:49 15.76459 -22.37979
#> 231  2 2015-09-10 07:01:50 15.76460 -22.37980
#> 232  2 2015-09-10 07:01:51 15.76461 -22.37980
#> 233  2 2015-09-10 07:01:52 15.76461 -22.37981
#> 234  2 2015-09-10 07:01:53 15.76461 -22.37981
#> 235  2 2015-09-10 07:01:54 15.76462 -22.37981
#> 236  2 2015-09-10 07:01:55 15.76463 -22.37982
#> 237  2 2015-09-10 07:01:56 15.76463 -22.37982
#> 238  2 2015-09-10 07:01:57 15.76464 -22.37982
#> 239  2 2015-09-10 07:01:58 15.76465 -22.37982
#> 240  2 2015-09-10 07:01:59 15.76466 -22.37982