Access/Modify the Projection of a Track Table

Functions to access or modify the projection of a data table. Changing the projection will trigger automatically the conversion of the locations in the new coordinate system.

projection(x)

projection(x) <- value

project(x, value)

Arguments

x
A track table.

value
A character string or a sp::CRS object representing the projection of the coordinates. "+proj=longlat" is suitable for the outputs of most GPS trackers.

Value

A track table.

Note

It is not possible to modify the projection if missing coordinates are present.

Author

Simon Garnier, garnier@njit.edu

Examples

data(tracks)

projection(tracks)

#> CRS arguments: +proj=longlat +ellps=WGS84 

tracks_projected <- project(tracks, "+proj=somerc")

#> Warning: NULL source CRS comment, falling back to PROJ string

#> Warning: CRS object has no comment

projection(tracks_projected)

#> CRS arguments:
#>  +proj=somerc +lat_0=0 +lon_0=0 +k_0=1 +x_0=0 +y_0=0 +datum=WGS84
#> +units=m +no_defs 

projection(tracks_projected) <- "+proj=longlat"
projection(tracks_projected)

#> CRS arguments: +proj=longlat +datum=WGS84 +no_defs