Extract or Replace Parts of a Track Table

Accessing columns, rows, or cells via $, [[, or [ is mostly similar to regular data frames. However, the behavior is sometimes different for track tables based on tibble and data.table. For more info, refer to tibble's and data.table's subsetting documentation.

# S3 method for track
[(x, ...)

# S3 method for track
[(x, ...) <- value

Arguments

x
A track table.

...
Other parameters to be passed to the extracting/subsetting functions of data.frame, tibble, and data.table.

value
A suitable replacement value: it will be repeated a whole number of times if necessary and it may be coerced: see the `Coercion` section in data.frame. If `NULL`, deletes the column if a single column is selected.

See also

Examples

library(trackdf)
t_df <- track_df(x = tracks$x, y = tracks$y, t = tracks$t, id = tracks$id,
  proj = "+proj=longlat", tz = "Africa/Windhoek")

tracks[1]

#> # A tibble: 7,194 x 1
#>    id   
#>    <chr>
#>  1 1    
#>  2 1    
#>  3 1    
#>  4 1    
#>  5 1    
#>  6 1    
#>  7 1    
#>  8 1    
#>  9 1    
#> 10 1    
#> # … with 7,184 more rows

tracks[1, ]

#> # A tibble: 1 x 4
#>   id    t                       x     y
#>   <chr> <dttm>              <dbl> <dbl>
#> 1 1     2015-09-10 07:00:00  15.8 -22.4

tracks[1, 1]

#> # A tibble: 1 x 1
#>   id   
#>   <chr>
#> 1 1    

tracks$id[tracks$id == "1"] <- "0"
tracks[tracks[, 1] == "0", 1] <- "1"