Convert a Track Table to/from Other Formats

The following methods will convert track tables to and from other common formats used for processing tracking and spatial data.

Usage

as_track(x, table = "df", ...)

# S3 method for MoveStack
as_track(x, table = "df", ...)

# S3 method for Move
as_track(x, table = "df", ...)

as_move(x, ...)

# S3 method for track
as_move(x, ...)

# S3 method for SpatialPointsDataFrame
as_track(x, table = "df", ...)

as_sp(x, ...)

# S3 method for track
as_sp(x, ...)

# S3 method for ltraj
as_track(x, table = "df", ...)

as_ltraj(x, ...)

# S3 method for track
as_ltraj(x, ...)

# S3 method for telemetry
as_track(x, table = "df", ...)

# S3 method for list
as_track(x, table = "df", ...)

as_telemetry(x, ...)

# S3 method for track
as_telemetry(x, ...)

# S3 method for moveData
as_track(x, table = "df", type = c("LL", "UTM"), ...)

as_moveHMM(x, ...)

# S3 method for track
as_moveHMM(x, ...)

Arguments

x

An object to convert.

table

A string indicating the class of the table on which the track table should be built. It can be a data.frame ("df", the default), a tibble ("tbl"), or a data.table ("dt").

...

Other parameters to be passed to:

track_df, track_tbl or track_dt if as_track is used.

moveVis::df2move if as_move is used.

sp::SpatialPointsDataFrame if as_sp is used.

adehabitatLT::as.ltraj if as_ltraj is used.

ctmm::as.telemetry if as_telemetry is used.

moveHMM::prepData if as_moveHMM is used.

type

For converting moveHMM::moveData to track table only, a character string indicating the type of coordinates stored in the moveHMM::moveData object: "LL" if longitude/latitude (default), "UTM" if easting/northing.

Value

The coordinates converted in the chosen format.

See also

Author

Simon Garnier, garnier@njit.edu

Examples

data(short_tracks)

if (requireNamespace("moveVis", quietly = TRUE)) {
  mv <- as_move(short_tracks)
  as_track(mv)
}
#> Track table [120 observations]
#> Number of tracks:  2 
#> Dimensions:  2D 
#> Geographic:  TRUE 
#> Projection:  +proj=longlat +ellps=WGS84 +no_defs 
#> Table class:  data frame
#>     id                   t        x         y
#> 1   X1 2015-09-10 07:01:00 15.76468 -22.37961
#> 2   X1 2015-09-10 07:01:01 15.76469 -22.37960
#> 3   X1 2015-09-10 07:01:02 15.76469 -22.37960
#> 4   X1 2015-09-10 07:01:03 15.76469 -22.37960
#> 5   X1 2015-09-10 07:01:04 15.76470 -22.37960
#> 6   X1 2015-09-10 07:01:05 15.76469 -22.37960
#> 7   X1 2015-09-10 07:01:06 15.76469 -22.37960
#> 8   X1 2015-09-10 07:01:07 15.76469 -22.37959
#> 9   X1 2015-09-10 07:01:08 15.76469 -22.37959
#> 10  X1 2015-09-10 07:01:09 15.76469 -22.37959
#> 11  X1 2015-09-10 07:01:10 15.76469 -22.37959
#> 12  X1 2015-09-10 07:01:11 15.76469 -22.37960
#> 13  X1 2015-09-10 07:01:12 15.76469 -22.37960
#> 14  X1 2015-09-10 07:01:13 15.76469 -22.37960
#> 15  X1 2015-09-10 07:01:14 15.76469 -22.37960
#> 16  X1 2015-09-10 07:01:15 15.76468 -22.37960
#> 17  X1 2015-09-10 07:01:16 15.76468 -22.37960
#> 18  X1 2015-09-10 07:01:17 15.76468 -22.37960
#> 19  X1 2015-09-10 07:01:18 15.76468 -22.37960
#> 20  X1 2015-09-10 07:01:19 15.76468 -22.37960
#> 21  X1 2015-09-10 07:01:20 15.76468 -22.37960
#> 22  X1 2015-09-10 07:01:21 15.76468 -22.37960
#> 23  X1 2015-09-10 07:01:22 15.76468 -22.37960
#> 24  X1 2015-09-10 07:01:23 15.76468 -22.37960
#> 25  X1 2015-09-10 07:01:24 15.76468 -22.37960
#> 26  X1 2015-09-10 07:01:25 15.76468 -22.37960
#> 27  X1 2015-09-10 07:01:26 15.76468 -22.37960
#> 28  X1 2015-09-10 07:01:27 15.76468 -22.37960
#> 29  X1 2015-09-10 07:01:28 15.76468 -22.37960
#> 30  X1 2015-09-10 07:01:29 15.76468 -22.37960
#> 31  X1 2015-09-10 07:01:30 15.76468 -22.37960
#> 32  X1 2015-09-10 07:01:31 15.76468 -22.37960
#> 33  X1 2015-09-10 07:01:32 15.76469 -22.37960
#> 34  X1 2015-09-10 07:01:33 15.76469 -22.37960
#> 35  X1 2015-09-10 07:01:34 15.76469 -22.37960
#> 36  X1 2015-09-10 07:01:35 15.76469 -22.37960
#> 37  X1 2015-09-10 07:01:36 15.76469 -22.37960
#> 38  X1 2015-09-10 07:01:37 15.76469 -22.37960
#> 39  X1 2015-09-10 07:01:38 15.76469 -22.37960
#> 40  X1 2015-09-10 07:01:39 15.76469 -22.37960
#> 41  X1 2015-09-10 07:01:40 15.76469 -22.37960
#> 42  X1 2015-09-10 07:01:41 15.76469 -22.37960
#> 43  X1 2015-09-10 07:01:42 15.76469 -22.37960
#> 44  X1 2015-09-10 07:01:43 15.76469 -22.37960
#> 45  X1 2015-09-10 07:01:44 15.76469 -22.37960
#> 46  X1 2015-09-10 07:01:45 15.76469 -22.37960
#> 47  X1 2015-09-10 07:01:46 15.76469 -22.37960
#> 48  X1 2015-09-10 07:01:47 15.76469 -22.37960
#> 49  X1 2015-09-10 07:01:48 15.76469 -22.37960
#> 50  X1 2015-09-10 07:01:49 15.76469 -22.37959
#> 51  X1 2015-09-10 07:01:50 15.76469 -22.37959
#> 52  X1 2015-09-10 07:01:51 15.76470 -22.37959
#> 53  X1 2015-09-10 07:01:52 15.76470 -22.37959
#> 54  X1 2015-09-10 07:01:53 15.76470 -22.37959
#> 55  X1 2015-09-10 07:01:54 15.76470 -22.37960
#> 56  X1 2015-09-10 07:01:55 15.76470 -22.37960
#> 57  X1 2015-09-10 07:01:56 15.76471 -22.37960
#> 58  X1 2015-09-10 07:01:57 15.76471 -22.37959
#> 59  X1 2015-09-10 07:01:58 15.76471 -22.37959
#> 60  X1 2015-09-10 07:01:59 15.76471 -22.37959
#> 61  X2 2015-09-10 07:01:00 15.76459 -22.37971
#> 62  X2 2015-09-10 07:01:01 15.76459 -22.37971
#> 63  X2 2015-09-10 07:01:02 15.76459 -22.37971
#> 64  X2 2015-09-10 07:01:03 15.76459 -22.37971
#> 65  X2 2015-09-10 07:01:04 15.76459 -22.37971
#> 66  X2 2015-09-10 07:01:05 15.76459 -22.37971
#> 67  X2 2015-09-10 07:01:06 15.76459 -22.37971
#> 68  X2 2015-09-10 07:01:07 15.76459 -22.37971
#> 69  X2 2015-09-10 07:01:08 15.76459 -22.37971
#> 70  X2 2015-09-10 07:01:09 15.76459 -22.37971
#> 71  X2 2015-09-10 07:01:10 15.76459 -22.37971
#> 72  X2 2015-09-10 07:01:11 15.76459 -22.37971
#> 73  X2 2015-09-10 07:01:12 15.76459 -22.37971
#> 74  X2 2015-09-10 07:01:13 15.76459 -22.37971
#> 75  X2 2015-09-10 07:01:14 15.76459 -22.37971
#> 76  X2 2015-09-10 07:01:15 15.76459 -22.37971
#> 77  X2 2015-09-10 07:01:16 15.76459 -22.37971
#> 78  X2 2015-09-10 07:01:17 15.76459 -22.37971
#> 79  X2 2015-09-10 07:01:18 15.76459 -22.37971
#> 80  X2 2015-09-10 07:01:19 15.76459 -22.37971
#> 81  X2 2015-09-10 07:01:20 15.76459 -22.37971
#> 82  X2 2015-09-10 07:01:21 15.76459 -22.37971
#> 83  X2 2015-09-10 07:01:22 15.76459 -22.37971
#> 84  X2 2015-09-10 07:01:23 15.76459 -22.37971
#> 85  X2 2015-09-10 07:01:24 15.76459 -22.37971
#> 86  X2 2015-09-10 07:01:25 15.76458 -22.37971
#> 87  X2 2015-09-10 07:01:26 15.76459 -22.37971
#> 88  X2 2015-09-10 07:01:27 15.76459 -22.37971
#> 89  X2 2015-09-10 07:01:28 15.76459 -22.37971
#> 90  X2 2015-09-10 07:01:29 15.76458 -22.37971
#> 91  X2 2015-09-10 07:01:30 15.76458 -22.37971
#> 92  X2 2015-09-10 07:01:31 15.76457 -22.37971
#> 93  X2 2015-09-10 07:01:32 15.76456 -22.37971
#> 94  X2 2015-09-10 07:01:33 15.76455 -22.37972
#> 95  X2 2015-09-10 07:01:34 15.76455 -22.37972
#> 96  X2 2015-09-10 07:01:35 15.76455 -22.37973
#> 97  X2 2015-09-10 07:01:36 15.76455 -22.37973
#> 98  X2 2015-09-10 07:01:37 15.76455 -22.37974
#> 99  X2 2015-09-10 07:01:38 15.76455 -22.37974
#> 100 X2 2015-09-10 07:01:39 15.76455 -22.37975
#> 101 X2 2015-09-10 07:01:40 15.76455 -22.37975
#> 102 X2 2015-09-10 07:01:41 15.76454 -22.37976
#> 103 X2 2015-09-10 07:01:42 15.76455 -22.37976
#> 104 X2 2015-09-10 07:01:43 15.76455 -22.37976
#> 105 X2 2015-09-10 07:01:44 15.76455 -22.37977
#> 106 X2 2015-09-10 07:01:45 15.76456 -22.37977
#> 107 X2 2015-09-10 07:01:46 15.76457 -22.37978
#> 108 X2 2015-09-10 07:01:47 15.76458 -22.37978
#> 109 X2 2015-09-10 07:01:48 15.76459 -22.37979
#> 110 X2 2015-09-10 07:01:49 15.76459 -22.37979
#> 111 X2 2015-09-10 07:01:50 15.76460 -22.37980
#> 112 X2 2015-09-10 07:01:51 15.76461 -22.37980
#> 113 X2 2015-09-10 07:01:52 15.76461 -22.37981
#> 114 X2 2015-09-10 07:01:53 15.76461 -22.37981
#> 115 X2 2015-09-10 07:01:54 15.76462 -22.37981
#> 116 X2 2015-09-10 07:01:55 15.76463 -22.37982
#> 117 X2 2015-09-10 07:01:56 15.76463 -22.37982
#> 118 X2 2015-09-10 07:01:57 15.76464 -22.37982
#> 119 X2 2015-09-10 07:01:58 15.76465 -22.37982
#> 120 X2 2015-09-10 07:01:59 15.76466 -22.37982

if (requireNamespace("sp", quietly = TRUE)) {
  sp <- as_sp(short_tracks)
  as_track(sp)
}
#> Track table [120 observations]
#> Number of tracks:  2 
#> Dimensions:  2D 
#> Geographic:  TRUE 
#> Projection:  +proj=longlat +ellps=WGS84 
#> Table class:  data frame
#>     id                   t        x         y
#> 1    1 2015-09-10 07:01:00 15.76468 -22.37961
#> 2    1 2015-09-10 07:01:01 15.76469 -22.37960
#> 3    1 2015-09-10 07:01:02 15.76469 -22.37960
#> 4    1 2015-09-10 07:01:03 15.76469 -22.37960
#> 5    1 2015-09-10 07:01:04 15.76470 -22.37960
#> 6    1 2015-09-10 07:01:05 15.76469 -22.37960
#> 7    1 2015-09-10 07:01:06 15.76469 -22.37960
#> 8    1 2015-09-10 07:01:07 15.76469 -22.37959
#> 9    1 2015-09-10 07:01:08 15.76469 -22.37959
#> 10   1 2015-09-10 07:01:09 15.76469 -22.37959
#> 11   1 2015-09-10 07:01:10 15.76469 -22.37959
#> 12   1 2015-09-10 07:01:11 15.76469 -22.37960
#> 13   1 2015-09-10 07:01:12 15.76469 -22.37960
#> 14   1 2015-09-10 07:01:13 15.76469 -22.37960
#> 15   1 2015-09-10 07:01:14 15.76469 -22.37960
#> 16   1 2015-09-10 07:01:15 15.76468 -22.37960
#> 17   1 2015-09-10 07:01:16 15.76468 -22.37960
#> 18   1 2015-09-10 07:01:17 15.76468 -22.37960
#> 19   1 2015-09-10 07:01:18 15.76468 -22.37960
#> 20   1 2015-09-10 07:01:19 15.76468 -22.37960
#> 21   1 2015-09-10 07:01:20 15.76468 -22.37960
#> 22   1 2015-09-10 07:01:21 15.76468 -22.37960
#> 23   1 2015-09-10 07:01:22 15.76468 -22.37960
#> 24   1 2015-09-10 07:01:23 15.76468 -22.37960
#> 25   1 2015-09-10 07:01:24 15.76468 -22.37960
#> 26   1 2015-09-10 07:01:25 15.76468 -22.37960
#> 27   1 2015-09-10 07:01:26 15.76468 -22.37960
#> 28   1 2015-09-10 07:01:27 15.76468 -22.37960
#> 29   1 2015-09-10 07:01:28 15.76468 -22.37960
#> 30   1 2015-09-10 07:01:29 15.76468 -22.37960
#> 31   1 2015-09-10 07:01:30 15.76468 -22.37960
#> 32   1 2015-09-10 07:01:31 15.76468 -22.37960
#> 33   1 2015-09-10 07:01:32 15.76469 -22.37960
#> 34   1 2015-09-10 07:01:33 15.76469 -22.37960
#> 35   1 2015-09-10 07:01:34 15.76469 -22.37960
#> 36   1 2015-09-10 07:01:35 15.76469 -22.37960
#> 37   1 2015-09-10 07:01:36 15.76469 -22.37960
#> 38   1 2015-09-10 07:01:37 15.76469 -22.37960
#> 39   1 2015-09-10 07:01:38 15.76469 -22.37960
#> 40   1 2015-09-10 07:01:39 15.76469 -22.37960
#> 41   1 2015-09-10 07:01:40 15.76469 -22.37960
#> 42   1 2015-09-10 07:01:41 15.76469 -22.37960
#> 43   1 2015-09-10 07:01:42 15.76469 -22.37960
#> 44   1 2015-09-10 07:01:43 15.76469 -22.37960
#> 45   1 2015-09-10 07:01:44 15.76469 -22.37960
#> 46   1 2015-09-10 07:01:45 15.76469 -22.37960
#> 47   1 2015-09-10 07:01:46 15.76469 -22.37960
#> 48   1 2015-09-10 07:01:47 15.76469 -22.37960
#> 49   1 2015-09-10 07:01:48 15.76469 -22.37960
#> 50   1 2015-09-10 07:01:49 15.76469 -22.37959
#> 51   1 2015-09-10 07:01:50 15.76469 -22.37959
#> 52   1 2015-09-10 07:01:51 15.76470 -22.37959
#> 53   1 2015-09-10 07:01:52 15.76470 -22.37959
#> 54   1 2015-09-10 07:01:53 15.76470 -22.37959
#> 55   1 2015-09-10 07:01:54 15.76470 -22.37960
#> 56   1 2015-09-10 07:01:55 15.76470 -22.37960
#> 57   1 2015-09-10 07:01:56 15.76471 -22.37960
#> 58   1 2015-09-10 07:01:57 15.76471 -22.37959
#> 59   1 2015-09-10 07:01:58 15.76471 -22.37959
#> 60   1 2015-09-10 07:01:59 15.76471 -22.37959
#> 61   2 2015-09-10 07:01:00 15.76459 -22.37971
#> 62   2 2015-09-10 07:01:01 15.76459 -22.37971
#> 63   2 2015-09-10 07:01:02 15.76459 -22.37971
#> 64   2 2015-09-10 07:01:03 15.76459 -22.37971
#> 65   2 2015-09-10 07:01:04 15.76459 -22.37971
#> 66   2 2015-09-10 07:01:05 15.76459 -22.37971
#> 67   2 2015-09-10 07:01:06 15.76459 -22.37971
#> 68   2 2015-09-10 07:01:07 15.76459 -22.37971
#> 69   2 2015-09-10 07:01:08 15.76459 -22.37971
#> 70   2 2015-09-10 07:01:09 15.76459 -22.37971
#> 71   2 2015-09-10 07:01:10 15.76459 -22.37971
#> 72   2 2015-09-10 07:01:11 15.76459 -22.37971
#> 73   2 2015-09-10 07:01:12 15.76459 -22.37971
#> 74   2 2015-09-10 07:01:13 15.76459 -22.37971
#> 75   2 2015-09-10 07:01:14 15.76459 -22.37971
#> 76   2 2015-09-10 07:01:15 15.76459 -22.37971
#> 77   2 2015-09-10 07:01:16 15.76459 -22.37971
#> 78   2 2015-09-10 07:01:17 15.76459 -22.37971
#> 79   2 2015-09-10 07:01:18 15.76459 -22.37971
#> 80   2 2015-09-10 07:01:19 15.76459 -22.37971
#> 81   2 2015-09-10 07:01:20 15.76459 -22.37971
#> 82   2 2015-09-10 07:01:21 15.76459 -22.37971
#> 83   2 2015-09-10 07:01:22 15.76459 -22.37971
#> 84   2 2015-09-10 07:01:23 15.76459 -22.37971
#> 85   2 2015-09-10 07:01:24 15.76459 -22.37971
#> 86   2 2015-09-10 07:01:25 15.76458 -22.37971
#> 87   2 2015-09-10 07:01:26 15.76459 -22.37971
#> 88   2 2015-09-10 07:01:27 15.76459 -22.37971
#> 89   2 2015-09-10 07:01:28 15.76459 -22.37971
#> 90   2 2015-09-10 07:01:29 15.76458 -22.37971
#> 91   2 2015-09-10 07:01:30 15.76458 -22.37971
#> 92   2 2015-09-10 07:01:31 15.76457 -22.37971
#> 93   2 2015-09-10 07:01:32 15.76456 -22.37971
#> 94   2 2015-09-10 07:01:33 15.76455 -22.37972
#> 95   2 2015-09-10 07:01:34 15.76455 -22.37972
#> 96   2 2015-09-10 07:01:35 15.76455 -22.37973
#> 97   2 2015-09-10 07:01:36 15.76455 -22.37973
#> 98   2 2015-09-10 07:01:37 15.76455 -22.37974
#> 99   2 2015-09-10 07:01:38 15.76455 -22.37974
#> 100  2 2015-09-10 07:01:39 15.76455 -22.37975
#> 101  2 2015-09-10 07:01:40 15.76455 -22.37975
#> 102  2 2015-09-10 07:01:41 15.76454 -22.37976
#> 103  2 2015-09-10 07:01:42 15.76455 -22.37976
#> 104  2 2015-09-10 07:01:43 15.76455 -22.37976
#> 105  2 2015-09-10 07:01:44 15.76455 -22.37977
#> 106  2 2015-09-10 07:01:45 15.76456 -22.37977
#> 107  2 2015-09-10 07:01:46 15.76457 -22.37978
#> 108  2 2015-09-10 07:01:47 15.76458 -22.37978
#> 109  2 2015-09-10 07:01:48 15.76459 -22.37979
#> 110  2 2015-09-10 07:01:49 15.76459 -22.37979
#> 111  2 2015-09-10 07:01:50 15.76460 -22.37980
#> 112  2 2015-09-10 07:01:51 15.76461 -22.37980
#> 113  2 2015-09-10 07:01:52 15.76461 -22.37981
#> 114  2 2015-09-10 07:01:53 15.76461 -22.37981
#> 115  2 2015-09-10 07:01:54 15.76462 -22.37981
#> 116  2 2015-09-10 07:01:55 15.76463 -22.37982
#> 117  2 2015-09-10 07:01:56 15.76463 -22.37982
#> 118  2 2015-09-10 07:01:57 15.76464 -22.37982
#> 119  2 2015-09-10 07:01:58 15.76465 -22.37982
#> 120  2 2015-09-10 07:01:59 15.76466 -22.37982

if (requireNamespace("adehabitatLT", quietly = TRUE)) {
  lt <- as_ltraj(short_tracks)
  as_track(lt)
}
#> Registered S3 methods overwritten by 'adehabitatMA':
#>   method                       from
#>   print.SpatialPixelsDataFrame sp  
#>   print.SpatialPixels          sp  
#> Track table [120 observations]
#> Number of tracks:  2 
#> Dimensions:  2D 
#> Geographic:  TRUE 
#> Projection:  +proj=longlat +ellps=WGS84 
#> Table class:  data frame
#>     id                   t        x         y
#> 1    1 2015-09-10 07:01:00 15.76468 -22.37961
#> 2    1 2015-09-10 07:01:01 15.76469 -22.37960
#> 3    1 2015-09-10 07:01:02 15.76469 -22.37960
#> 4    1 2015-09-10 07:01:03 15.76469 -22.37960
#> 5    1 2015-09-10 07:01:04 15.76470 -22.37960
#> 6    1 2015-09-10 07:01:05 15.76469 -22.37960
#> 7    1 2015-09-10 07:01:06 15.76469 -22.37960
#> 8    1 2015-09-10 07:01:07 15.76469 -22.37959
#> 9    1 2015-09-10 07:01:08 15.76469 -22.37959
#> 10   1 2015-09-10 07:01:09 15.76469 -22.37959
#> 11   1 2015-09-10 07:01:10 15.76469 -22.37959
#> 12   1 2015-09-10 07:01:11 15.76469 -22.37960
#> 13   1 2015-09-10 07:01:12 15.76469 -22.37960
#> 14   1 2015-09-10 07:01:13 15.76469 -22.37960
#> 15   1 2015-09-10 07:01:14 15.76469 -22.37960
#> 16   1 2015-09-10 07:01:15 15.76468 -22.37960
#> 17   1 2015-09-10 07:01:16 15.76468 -22.37960
#> 18   1 2015-09-10 07:01:17 15.76468 -22.37960
#> 19   1 2015-09-10 07:01:18 15.76468 -22.37960
#> 20   1 2015-09-10 07:01:19 15.76468 -22.37960
#> 21   1 2015-09-10 07:01:20 15.76468 -22.37960
#> 22   1 2015-09-10 07:01:21 15.76468 -22.37960
#> 23   1 2015-09-10 07:01:22 15.76468 -22.37960
#> 24   1 2015-09-10 07:01:23 15.76468 -22.37960
#> 25   1 2015-09-10 07:01:24 15.76468 -22.37960
#> 26   1 2015-09-10 07:01:25 15.76468 -22.37960
#> 27   1 2015-09-10 07:01:26 15.76468 -22.37960
#> 28   1 2015-09-10 07:01:27 15.76468 -22.37960
#> 29   1 2015-09-10 07:01:28 15.76468 -22.37960
#> 30   1 2015-09-10 07:01:29 15.76468 -22.37960
#> 31   1 2015-09-10 07:01:30 15.76468 -22.37960
#> 32   1 2015-09-10 07:01:31 15.76468 -22.37960
#> 33   1 2015-09-10 07:01:32 15.76469 -22.37960
#> 34   1 2015-09-10 07:01:33 15.76469 -22.37960
#> 35   1 2015-09-10 07:01:34 15.76469 -22.37960
#> 36   1 2015-09-10 07:01:35 15.76469 -22.37960
#> 37   1 2015-09-10 07:01:36 15.76469 -22.37960
#> 38   1 2015-09-10 07:01:37 15.76469 -22.37960
#> 39   1 2015-09-10 07:01:38 15.76469 -22.37960
#> 40   1 2015-09-10 07:01:39 15.76469 -22.37960
#> 41   1 2015-09-10 07:01:40 15.76469 -22.37960
#> 42   1 2015-09-10 07:01:41 15.76469 -22.37960
#> 43   1 2015-09-10 07:01:42 15.76469 -22.37960
#> 44   1 2015-09-10 07:01:43 15.76469 -22.37960
#> 45   1 2015-09-10 07:01:44 15.76469 -22.37960
#> 46   1 2015-09-10 07:01:45 15.76469 -22.37960
#> 47   1 2015-09-10 07:01:46 15.76469 -22.37960
#> 48   1 2015-09-10 07:01:47 15.76469 -22.37960
#> 49   1 2015-09-10 07:01:48 15.76469 -22.37960
#> 50   1 2015-09-10 07:01:49 15.76469 -22.37959
#> 51   1 2015-09-10 07:01:50 15.76469 -22.37959
#> 52   1 2015-09-10 07:01:51 15.76470 -22.37959
#> 53   1 2015-09-10 07:01:52 15.76470 -22.37959
#> 54   1 2015-09-10 07:01:53 15.76470 -22.37959
#> 55   1 2015-09-10 07:01:54 15.76470 -22.37960
#> 56   1 2015-09-10 07:01:55 15.76470 -22.37960
#> 57   1 2015-09-10 07:01:56 15.76471 -22.37960
#> 58   1 2015-09-10 07:01:57 15.76471 -22.37959
#> 59   1 2015-09-10 07:01:58 15.76471 -22.37959
#> 60   1 2015-09-10 07:01:59 15.76471 -22.37959
#> 61   2 2015-09-10 07:01:00 15.76459 -22.37971
#> 62   2 2015-09-10 07:01:01 15.76459 -22.37971
#> 63   2 2015-09-10 07:01:02 15.76459 -22.37971
#> 64   2 2015-09-10 07:01:03 15.76459 -22.37971
#> 65   2 2015-09-10 07:01:04 15.76459 -22.37971
#> 66   2 2015-09-10 07:01:05 15.76459 -22.37971
#> 67   2 2015-09-10 07:01:06 15.76459 -22.37971
#> 68   2 2015-09-10 07:01:07 15.76459 -22.37971
#> 69   2 2015-09-10 07:01:08 15.76459 -22.37971
#> 70   2 2015-09-10 07:01:09 15.76459 -22.37971
#> 71   2 2015-09-10 07:01:10 15.76459 -22.37971
#> 72   2 2015-09-10 07:01:11 15.76459 -22.37971
#> 73   2 2015-09-10 07:01:12 15.76459 -22.37971
#> 74   2 2015-09-10 07:01:13 15.76459 -22.37971
#> 75   2 2015-09-10 07:01:14 15.76459 -22.37971
#> 76   2 2015-09-10 07:01:15 15.76459 -22.37971
#> 77   2 2015-09-10 07:01:16 15.76459 -22.37971
#> 78   2 2015-09-10 07:01:17 15.76459 -22.37971
#> 79   2 2015-09-10 07:01:18 15.76459 -22.37971
#> 80   2 2015-09-10 07:01:19 15.76459 -22.37971
#> 81   2 2015-09-10 07:01:20 15.76459 -22.37971
#> 82   2 2015-09-10 07:01:21 15.76459 -22.37971
#> 83   2 2015-09-10 07:01:22 15.76459 -22.37971
#> 84   2 2015-09-10 07:01:23 15.76459 -22.37971
#> 85   2 2015-09-10 07:01:24 15.76459 -22.37971
#> 86   2 2015-09-10 07:01:25 15.76458 -22.37971
#> 87   2 2015-09-10 07:01:26 15.76459 -22.37971
#> 88   2 2015-09-10 07:01:27 15.76459 -22.37971
#> 89   2 2015-09-10 07:01:28 15.76459 -22.37971
#> 90   2 2015-09-10 07:01:29 15.76458 -22.37971
#> 91   2 2015-09-10 07:01:30 15.76458 -22.37971
#> 92   2 2015-09-10 07:01:31 15.76457 -22.37971
#> 93   2 2015-09-10 07:01:32 15.76456 -22.37971
#> 94   2 2015-09-10 07:01:33 15.76455 -22.37972
#> 95   2 2015-09-10 07:01:34 15.76455 -22.37972
#> 96   2 2015-09-10 07:01:35 15.76455 -22.37973
#> 97   2 2015-09-10 07:01:36 15.76455 -22.37973
#> 98   2 2015-09-10 07:01:37 15.76455 -22.37974
#> 99   2 2015-09-10 07:01:38 15.76455 -22.37974
#> 100  2 2015-09-10 07:01:39 15.76455 -22.37975
#> 101  2 2015-09-10 07:01:40 15.76455 -22.37975
#> 102  2 2015-09-10 07:01:41 15.76454 -22.37976
#> 103  2 2015-09-10 07:01:42 15.76455 -22.37976
#> 104  2 2015-09-10 07:01:43 15.76455 -22.37976
#> 105  2 2015-09-10 07:01:44 15.76455 -22.37977
#> 106  2 2015-09-10 07:01:45 15.76456 -22.37977
#> 107  2 2015-09-10 07:01:46 15.76457 -22.37978
#> 108  2 2015-09-10 07:01:47 15.76458 -22.37978
#> 109  2 2015-09-10 07:01:48 15.76459 -22.37979
#> 110  2 2015-09-10 07:01:49 15.76459 -22.37979
#> 111  2 2015-09-10 07:01:50 15.76460 -22.37980
#> 112  2 2015-09-10 07:01:51 15.76461 -22.37980
#> 113  2 2015-09-10 07:01:52 15.76461 -22.37981
#> 114  2 2015-09-10 07:01:53 15.76461 -22.37981
#> 115  2 2015-09-10 07:01:54 15.76462 -22.37981
#> 116  2 2015-09-10 07:01:55 15.76463 -22.37982
#> 117  2 2015-09-10 07:01:56 15.76463 -22.37982
#> 118  2 2015-09-10 07:01:57 15.76464 -22.37982
#> 119  2 2015-09-10 07:01:58 15.76465 -22.37982
#> 120  2 2015-09-10 07:01:59 15.76466 -22.37982

if (requireNamespace("ctmm", quietly = TRUE)) {
  tl <- as_telemetry(short_tracks)
  as_track(tl)
}
#> Minimum sampling interval of 1 seconds in 1
#> Minimum sampling interval of 1 seconds in 2
#> Warning: Unrecognized time zone 'CAT'
#> Warning: Unrecognized time zone 'CAT'
#> Track table [120 observations]
#> Number of tracks:  2 
#> Dimensions:  2D 
#> Geographic:  TRUE 
#> Projection:  +proj=tpeqd +lat_1=-22.3797489764867 +lon_1=15.7645721041225 +lat_2=-22.3796046594646 +lon_2=15.7646744509934 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs 
#> Table class:  data frame
#> Warning: unknown timezone 'CAT'
#>     id                   t        x         y
#> 1    1 2015-09-10 05:01:00 15.76468 -22.37961
#> 2    1 2015-09-10 05:01:01 15.76469 -22.37960
#> 3    1 2015-09-10 05:01:02 15.76469 -22.37960
#> 4    1 2015-09-10 05:01:03 15.76469 -22.37960
#> 5    1 2015-09-10 05:01:04 15.76470 -22.37960
#> 6    1 2015-09-10 05:01:05 15.76469 -22.37960
#> 7    1 2015-09-10 05:01:06 15.76469 -22.37960
#> 8    1 2015-09-10 05:01:07 15.76469 -22.37959
#> 9    1 2015-09-10 05:01:08 15.76469 -22.37959
#> 10   1 2015-09-10 05:01:09 15.76469 -22.37959
#> 11   1 2015-09-10 05:01:10 15.76469 -22.37959
#> 12   1 2015-09-10 05:01:11 15.76469 -22.37960
#> 13   1 2015-09-10 05:01:12 15.76469 -22.37960
#> 14   1 2015-09-10 05:01:13 15.76469 -22.37960
#> 15   1 2015-09-10 05:01:14 15.76469 -22.37960
#> 16   1 2015-09-10 05:01:15 15.76468 -22.37960
#> 17   1 2015-09-10 05:01:16 15.76468 -22.37960
#> 18   1 2015-09-10 05:01:17 15.76468 -22.37960
#> 19   1 2015-09-10 05:01:18 15.76468 -22.37960
#> 20   1 2015-09-10 05:01:19 15.76468 -22.37960
#> 21   1 2015-09-10 05:01:20 15.76468 -22.37960
#> 22   1 2015-09-10 05:01:21 15.76468 -22.37960
#> 23   1 2015-09-10 05:01:22 15.76468 -22.37960
#> 24   1 2015-09-10 05:01:23 15.76468 -22.37960
#> 25   1 2015-09-10 05:01:24 15.76468 -22.37960
#> 26   1 2015-09-10 05:01:25 15.76468 -22.37960
#> 27   1 2015-09-10 05:01:26 15.76468 -22.37960
#> 28   1 2015-09-10 05:01:27 15.76468 -22.37960
#> 29   1 2015-09-10 05:01:28 15.76468 -22.37960
#> 30   1 2015-09-10 05:01:29 15.76468 -22.37960
#> 31   1 2015-09-10 05:01:30 15.76468 -22.37960
#> 32   1 2015-09-10 05:01:31 15.76468 -22.37960
#> 33   1 2015-09-10 05:01:32 15.76469 -22.37960
#> 34   1 2015-09-10 05:01:33 15.76469 -22.37960
#> 35   1 2015-09-10 05:01:34 15.76469 -22.37960
#> 36   1 2015-09-10 05:01:35 15.76469 -22.37960
#> 37   1 2015-09-10 05:01:36 15.76469 -22.37960
#> 38   1 2015-09-10 05:01:37 15.76469 -22.37960
#> 39   1 2015-09-10 05:01:38 15.76469 -22.37960
#> 40   1 2015-09-10 05:01:39 15.76469 -22.37960
#> 41   1 2015-09-10 05:01:40 15.76469 -22.37960
#> 42   1 2015-09-10 05:01:41 15.76469 -22.37960
#> 43   1 2015-09-10 05:01:42 15.76469 -22.37960
#> 44   1 2015-09-10 05:01:43 15.76469 -22.37960
#> 45   1 2015-09-10 05:01:44 15.76469 -22.37960
#> 46   1 2015-09-10 05:01:45 15.76469 -22.37960
#> 47   1 2015-09-10 05:01:46 15.76469 -22.37960
#> 48   1 2015-09-10 05:01:47 15.76469 -22.37960
#> 49   1 2015-09-10 05:01:48 15.76469 -22.37960
#> 50   1 2015-09-10 05:01:49 15.76469 -22.37959
#> 51   1 2015-09-10 05:01:50 15.76469 -22.37959
#> 52   1 2015-09-10 05:01:51 15.76470 -22.37959
#> 53   1 2015-09-10 05:01:52 15.76470 -22.37959
#> 54   1 2015-09-10 05:01:53 15.76470 -22.37959
#> 55   1 2015-09-10 05:01:54 15.76470 -22.37960
#> 56   1 2015-09-10 05:01:55 15.76470 -22.37960
#> 57   1 2015-09-10 05:01:56 15.76471 -22.37960
#> 58   1 2015-09-10 05:01:57 15.76471 -22.37959
#> 59   1 2015-09-10 05:01:58 15.76471 -22.37959
#> 60   1 2015-09-10 05:01:59 15.76471 -22.37959
#> 61   2 2015-09-10 05:01:00 15.76459 -22.37971
#> 62   2 2015-09-10 05:01:01 15.76459 -22.37971
#> 63   2 2015-09-10 05:01:02 15.76459 -22.37971
#> 64   2 2015-09-10 05:01:03 15.76459 -22.37971
#> 65   2 2015-09-10 05:01:04 15.76459 -22.37971
#> 66   2 2015-09-10 05:01:05 15.76459 -22.37971
#> 67   2 2015-09-10 05:01:06 15.76459 -22.37971
#> 68   2 2015-09-10 05:01:07 15.76459 -22.37971
#> 69   2 2015-09-10 05:01:08 15.76459 -22.37971
#> 70   2 2015-09-10 05:01:09 15.76459 -22.37971
#> 71   2 2015-09-10 05:01:10 15.76459 -22.37971
#> 72   2 2015-09-10 05:01:11 15.76459 -22.37971
#> 73   2 2015-09-10 05:01:12 15.76459 -22.37971
#> 74   2 2015-09-10 05:01:13 15.76459 -22.37971
#> 75   2 2015-09-10 05:01:14 15.76459 -22.37971
#> 76   2 2015-09-10 05:01:15 15.76459 -22.37971
#> 77   2 2015-09-10 05:01:16 15.76459 -22.37971
#> 78   2 2015-09-10 05:01:17 15.76459 -22.37971
#> 79   2 2015-09-10 05:01:18 15.76459 -22.37971
#> 80   2 2015-09-10 05:01:19 15.76459 -22.37971
#> 81   2 2015-09-10 05:01:20 15.76459 -22.37971
#> 82   2 2015-09-10 05:01:21 15.76459 -22.37971
#> 83   2 2015-09-10 05:01:22 15.76459 -22.37971
#> 84   2 2015-09-10 05:01:23 15.76459 -22.37971
#> 85   2 2015-09-10 05:01:24 15.76459 -22.37971
#> 86   2 2015-09-10 05:01:25 15.76458 -22.37971
#> 87   2 2015-09-10 05:01:26 15.76459 -22.37971
#> 88   2 2015-09-10 05:01:27 15.76459 -22.37971
#> 89   2 2015-09-10 05:01:28 15.76459 -22.37971
#> 90   2 2015-09-10 05:01:29 15.76458 -22.37971
#> 91   2 2015-09-10 05:01:30 15.76458 -22.37971
#> 92   2 2015-09-10 05:01:31 15.76457 -22.37971
#> 93   2 2015-09-10 05:01:32 15.76456 -22.37971
#> 94   2 2015-09-10 05:01:33 15.76455 -22.37972
#> 95   2 2015-09-10 05:01:34 15.76455 -22.37972
#> 96   2 2015-09-10 05:01:35 15.76455 -22.37973
#> 97   2 2015-09-10 05:01:36 15.76455 -22.37973
#> 98   2 2015-09-10 05:01:37 15.76455 -22.37974
#> 99   2 2015-09-10 05:01:38 15.76455 -22.37974
#> 100  2 2015-09-10 05:01:39 15.76455 -22.37975
#> 101  2 2015-09-10 05:01:40 15.76455 -22.37975
#> 102  2 2015-09-10 05:01:41 15.76454 -22.37976
#> 103  2 2015-09-10 05:01:42 15.76455 -22.37976
#> 104  2 2015-09-10 05:01:43 15.76455 -22.37976
#> 105  2 2015-09-10 05:01:44 15.76455 -22.37977
#> 106  2 2015-09-10 05:01:45 15.76456 -22.37977
#> 107  2 2015-09-10 05:01:46 15.76457 -22.37978
#> 108  2 2015-09-10 05:01:47 15.76458 -22.37978
#> 109  2 2015-09-10 05:01:48 15.76459 -22.37979
#> 110  2 2015-09-10 05:01:49 15.76459 -22.37979
#> 111  2 2015-09-10 05:01:50 15.76460 -22.37980
#> 112  2 2015-09-10 05:01:51 15.76461 -22.37980
#> 113  2 2015-09-10 05:01:52 15.76461 -22.37981
#> 114  2 2015-09-10 05:01:53 15.76461 -22.37981
#> 115  2 2015-09-10 05:01:54 15.76462 -22.37981
#> 116  2 2015-09-10 05:01:55 15.76463 -22.37982
#> 117  2 2015-09-10 05:01:56 15.76463 -22.37982
#> 118  2 2015-09-10 05:01:57 15.76464 -22.37982
#> 119  2 2015-09-10 05:01:58 15.76465 -22.37982
#> 120  2 2015-09-10 05:01:59 15.76466 -22.37982

if (requireNamespace("moveHMM", quietly = TRUE)) {
  hhm <- as_moveHMM(short_tracks, type = "LL")
  as_track(hhm)
}
#> Warning: NULL source CRS comment, falling back to PROJ string
#> Track table [120 observations]
#> Number of tracks:  2 
#> Dimensions:  2D 
#> Geographic:  TRUE 
#> Projection:  +proj=longlat +datum=WGS84 +no_defs 
#> Table class:  data frame
#>     id                   t        x         y
#> 1    1 2015-09-10 07:01:00 15.76468 -22.37961
#> 2    1 2015-09-10 07:01:01 15.76469 -22.37960
#> 3    1 2015-09-10 07:01:02 15.76469 -22.37960
#> 4    1 2015-09-10 07:01:03 15.76469 -22.37960
#> 5    1 2015-09-10 07:01:04 15.76470 -22.37960
#> 6    1 2015-09-10 07:01:05 15.76469 -22.37960
#> 7    1 2015-09-10 07:01:06 15.76469 -22.37960
#> 8    1 2015-09-10 07:01:07 15.76469 -22.37959
#> 9    1 2015-09-10 07:01:08 15.76469 -22.37959
#> 10   1 2015-09-10 07:01:09 15.76469 -22.37959
#> 11   1 2015-09-10 07:01:10 15.76469 -22.37959
#> 12   1 2015-09-10 07:01:11 15.76469 -22.37960
#> 13   1 2015-09-10 07:01:12 15.76469 -22.37960
#> 14   1 2015-09-10 07:01:13 15.76469 -22.37960
#> 15   1 2015-09-10 07:01:14 15.76469 -22.37960
#> 16   1 2015-09-10 07:01:15 15.76468 -22.37960
#> 17   1 2015-09-10 07:01:16 15.76468 -22.37960
#> 18   1 2015-09-10 07:01:17 15.76468 -22.37960
#> 19   1 2015-09-10 07:01:18 15.76468 -22.37960
#> 20   1 2015-09-10 07:01:19 15.76468 -22.37960
#> 21   1 2015-09-10 07:01:20 15.76468 -22.37960
#> 22   1 2015-09-10 07:01:21 15.76468 -22.37960
#> 23   1 2015-09-10 07:01:22 15.76468 -22.37960
#> 24   1 2015-09-10 07:01:23 15.76468 -22.37960
#> 25   1 2015-09-10 07:01:24 15.76468 -22.37960
#> 26   1 2015-09-10 07:01:25 15.76468 -22.37960
#> 27   1 2015-09-10 07:01:26 15.76468 -22.37960
#> 28   1 2015-09-10 07:01:27 15.76468 -22.37960
#> 29   1 2015-09-10 07:01:28 15.76468 -22.37960
#> 30   1 2015-09-10 07:01:29 15.76468 -22.37960
#> 31   1 2015-09-10 07:01:30 15.76468 -22.37960
#> 32   1 2015-09-10 07:01:31 15.76468 -22.37960
#> 33   1 2015-09-10 07:01:32 15.76469 -22.37960
#> 34   1 2015-09-10 07:01:33 15.76469 -22.37960
#> 35   1 2015-09-10 07:01:34 15.76469 -22.37960
#> 36   1 2015-09-10 07:01:35 15.76469 -22.37960
#> 37   1 2015-09-10 07:01:36 15.76469 -22.37960
#> 38   1 2015-09-10 07:01:37 15.76469 -22.37960
#> 39   1 2015-09-10 07:01:38 15.76469 -22.37960
#> 40   1 2015-09-10 07:01:39 15.76469 -22.37960
#> 41   1 2015-09-10 07:01:40 15.76469 -22.37960
#> 42   1 2015-09-10 07:01:41 15.76469 -22.37960
#> 43   1 2015-09-10 07:01:42 15.76469 -22.37960
#> 44   1 2015-09-10 07:01:43 15.76469 -22.37960
#> 45   1 2015-09-10 07:01:44 15.76469 -22.37960
#> 46   1 2015-09-10 07:01:45 15.76469 -22.37960
#> 47   1 2015-09-10 07:01:46 15.76469 -22.37960
#> 48   1 2015-09-10 07:01:47 15.76469 -22.37960
#> 49   1 2015-09-10 07:01:48 15.76469 -22.37960
#> 50   1 2015-09-10 07:01:49 15.76469 -22.37959
#> 51   1 2015-09-10 07:01:50 15.76469 -22.37959
#> 52   1 2015-09-10 07:01:51 15.76470 -22.37959
#> 53   1 2015-09-10 07:01:52 15.76470 -22.37959
#> 54   1 2015-09-10 07:01:53 15.76470 -22.37959
#> 55   1 2015-09-10 07:01:54 15.76470 -22.37960
#> 56   1 2015-09-10 07:01:55 15.76470 -22.37960
#> 57   1 2015-09-10 07:01:56 15.76471 -22.37960
#> 58   1 2015-09-10 07:01:57 15.76471 -22.37959
#> 59   1 2015-09-10 07:01:58 15.76471 -22.37959
#> 60   1 2015-09-10 07:01:59 15.76471 -22.37959
#> 61   2 2015-09-10 07:01:00 15.76459 -22.37971
#> 62   2 2015-09-10 07:01:01 15.76459 -22.37971
#> 63   2 2015-09-10 07:01:02 15.76459 -22.37971
#> 64   2 2015-09-10 07:01:03 15.76459 -22.37971
#> 65   2 2015-09-10 07:01:04 15.76459 -22.37971
#> 66   2 2015-09-10 07:01:05 15.76459 -22.37971
#> 67   2 2015-09-10 07:01:06 15.76459 -22.37971
#> 68   2 2015-09-10 07:01:07 15.76459 -22.37971
#> 69   2 2015-09-10 07:01:08 15.76459 -22.37971
#> 70   2 2015-09-10 07:01:09 15.76459 -22.37971
#> 71   2 2015-09-10 07:01:10 15.76459 -22.37971
#> 72   2 2015-09-10 07:01:11 15.76459 -22.37971
#> 73   2 2015-09-10 07:01:12 15.76459 -22.37971
#> 74   2 2015-09-10 07:01:13 15.76459 -22.37971
#> 75   2 2015-09-10 07:01:14 15.76459 -22.37971
#> 76   2 2015-09-10 07:01:15 15.76459 -22.37971
#> 77   2 2015-09-10 07:01:16 15.76459 -22.37971
#> 78   2 2015-09-10 07:01:17 15.76459 -22.37971
#> 79   2 2015-09-10 07:01:18 15.76459 -22.37971
#> 80   2 2015-09-10 07:01:19 15.76459 -22.37971
#> 81   2 2015-09-10 07:01:20 15.76459 -22.37971
#> 82   2 2015-09-10 07:01:21 15.76459 -22.37971
#> 83   2 2015-09-10 07:01:22 15.76459 -22.37971
#> 84   2 2015-09-10 07:01:23 15.76459 -22.37971
#> 85   2 2015-09-10 07:01:24 15.76459 -22.37971
#> 86   2 2015-09-10 07:01:25 15.76458 -22.37971
#> 87   2 2015-09-10 07:01:26 15.76459 -22.37971
#> 88   2 2015-09-10 07:01:27 15.76459 -22.37971
#> 89   2 2015-09-10 07:01:28 15.76459 -22.37971
#> 90   2 2015-09-10 07:01:29 15.76458 -22.37971
#> 91   2 2015-09-10 07:01:30 15.76458 -22.37971
#> 92   2 2015-09-10 07:01:31 15.76457 -22.37971
#> 93   2 2015-09-10 07:01:32 15.76456 -22.37971
#> 94   2 2015-09-10 07:01:33 15.76455 -22.37972
#> 95   2 2015-09-10 07:01:34 15.76455 -22.37972
#> 96   2 2015-09-10 07:01:35 15.76455 -22.37973
#> 97   2 2015-09-10 07:01:36 15.76455 -22.37973
#> 98   2 2015-09-10 07:01:37 15.76455 -22.37974
#> 99   2 2015-09-10 07:01:38 15.76455 -22.37974
#> 100  2 2015-09-10 07:01:39 15.76455 -22.37975
#> 101  2 2015-09-10 07:01:40 15.76455 -22.37975
#> 102  2 2015-09-10 07:01:41 15.76454 -22.37976
#> 103  2 2015-09-10 07:01:42 15.76455 -22.37976
#> 104  2 2015-09-10 07:01:43 15.76455 -22.37976
#> 105  2 2015-09-10 07:01:44 15.76455 -22.37977
#> 106  2 2015-09-10 07:01:45 15.76456 -22.37977
#> 107  2 2015-09-10 07:01:46 15.76457 -22.37978
#> 108  2 2015-09-10 07:01:47 15.76458 -22.37978
#> 109  2 2015-09-10 07:01:48 15.76459 -22.37979
#> 110  2 2015-09-10 07:01:49 15.76459 -22.37979
#> 111  2 2015-09-10 07:01:50 15.76460 -22.37980
#> 112  2 2015-09-10 07:01:51 15.76461 -22.37980
#> 113  2 2015-09-10 07:01:52 15.76461 -22.37981
#> 114  2 2015-09-10 07:01:53 15.76461 -22.37981
#> 115  2 2015-09-10 07:01:54 15.76462 -22.37981
#> 116  2 2015-09-10 07:01:55 15.76463 -22.37982
#> 117  2 2015-09-10 07:01:56 15.76463 -22.37982
#> 118  2 2015-09-10 07:01:57 15.76464 -22.37982
#> 119  2 2015-09-10 07:01:58 15.76465 -22.37982
#> 120  2 2015-09-10 07:01:59 15.76466 -22.37982