Extract or Replace Parts of a Track Table

Accessing columns, rows, or cells via $, [[, or [ is mostly similar to regular data frames. However, the behavior is sometimes different for track tables based on tibble and data.table. For more info, refer to tibble's and data.table's subsetting documentation.

# S3 method for track
[(x, ...)

# S3 method for track
[(x, ...) <- value

Arguments

x
A track table.

...
Other parameters to be passed to the extracting/subsetting functions of data.frame, tibble, and data.table.

value
A suitable replacement value: it will be repeated a whole number of times if necessary and it may be coerced: see the `Coercion` section in data.frame. If `NULL`, deletes the column if a single column is selected.

Value

A subset of the track table is [ is called, or a modified version of the track table if [<- is called.

See also

Examples

data(short_tracks)

short_tracks[1]

#>     id
#> 1    1
#> 2    1
#> 3    1
#> 4    1
#> 5    1
#> 6    1
#> 7    1
#> 8    1
#> 9    1
#> 10   1
#> 11   1
#> 12   1
#> 13   1
#> 14   1
#> 15   1
#> 16   1
#> 17   1
#> 18   1
#> 19   1
#> 20   1
#> 21   1
#> 22   1
#> 23   1
#> 24   1
#> 25   1
#> 26   1
#> 27   1
#> 28   1
#> 29   1
#> 30   1
#> 31   1
#> 32   1
#> 33   1
#> 34   1
#> 35   1
#> 36   1
#> 37   1
#> 38   1
#> 39   1
#> 40   1
#> 41   1
#> 42   1
#> 43   1
#> 44   1
#> 45   1
#> 46   1
#> 47   1
#> 48   1
#> 49   1
#> 50   1
#> 51   1
#> 52   1
#> 53   1
#> 54   1
#> 55   1
#> 56   1
#> 57   1
#> 58   1
#> 59   1
#> 60   1
#> 61   2
#> 62   2
#> 63   2
#> 64   2
#> 65   2
#> 66   2
#> 67   2
#> 68   2
#> 69   2
#> 70   2
#> 71   2
#> 72   2
#> 73   2
#> 74   2
#> 75   2
#> 76   2
#> 77   2
#> 78   2
#> 79   2
#> 80   2
#> 81   2
#> 82   2
#> 83   2
#> 84   2
#> 85   2
#> 86   2
#> 87   2
#> 88   2
#> 89   2
#> 90   2
#> 91   2
#> 92   2
#> 93   2
#> 94   2
#> 95   2
#> 96   2
#> 97   2
#> 98   2
#> 99   2
#> 100  2
#> 101  2
#> 102  2
#> 103  2
#> 104  2
#> 105  2
#> 106  2
#> 107  2
#> 108  2
#> 109  2
#> 110  2
#> 111  2
#> 112  2
#> 113  2
#> 114  2
#> 115  2
#> 116  2
#> 117  2
#> 118  2
#> 119  2
#> 120  2

short_tracks[1, ]

#> Track table [1 observations]
#> Number of tracks:  1 
#> Dimensions:  2D 
#> Geographic:  TRUE 
#> Projection:  +proj=longlat +ellps=WGS84 
#> Table class:  data frame
#>   id                   t        x         y
#> 1  1 2015-09-10 07:01:00 15.76468 -22.37961

short_tracks[1, 1]

#> [1] "1"

short_tracks$id[short_tracks$id == "1"] <- "0"
short_tracks[short_tracks[, 1] == "0", 1] <- "1"